# api-onboarding/src/doc/api/resolved.openapi_registration_v1.yaml
openapi: 3.1.0
info:
  title: API Onboarding - Registration API
  version: 1.0.0
  description: Minimal API surface for submitting Candidates to the registration REST controller.
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
  contact:
    name: API Onboarding Team
    email: me@proton.me
servers:
  - url: /api/v1
    description: API onboarding server running in the DEV cluster
paths:
  /registrations:
    post:
      operationId: submitCandidate
      summary: Submit a Candidate
      description: Accepts a Candidate payload describing a specification source to be processed.
      tags:
        - registrations
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Candidate'
            examples:
              valid-candidate:
                $ref: '#/components/examples/valid-candidate-uri'
              scored-candidate:
                $ref: '#/components/examples/scored-candidate-uri'
              registered-candidate:
                $ref: '#/components/examples/registered-candidate-uri'
              rejected-candidate:
                $ref: '#/components/examples/rejected-candidate-uri'
              invalid-candidate-missing-uri:
                $ref: '#/components/examples/invalid-candidate-missing-uri'
              invalid-candidate-malformed-uri:
                $ref: '#/components/examples/invalid-candidate-malformed-uri'
              invalid-candidate-not-found-uri:
                $ref: '#/components/examples/invalid-candidate-not-found-uri'
      responses:
        '200':
          description: Ok — candidate has been validated, and scored successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CandidateProcessed'
              examples:
                valid-candidate:
                  $ref: '#/components/examples/candidate-validated'
                scored-candidate:
                  $ref: '#/components/examples/candidate-scored'
        '201':
          description: Ok — candidate has been validated, scored, and registered successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CandidateProcessed'
              examples:
                registered-candidate:
                  $ref: '#/components/examples/candidate-registered'
        '400':
          description: The request was malformed or could not be processed.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              examples:
                invalid-candidate-missing-uri:
                  $ref: '#/components/examples/candidate-invalid-missing-uri'
                invalid-candidate-malformed-uri:
                  $ref: '#/components/examples/candidate-invalid-malformed-uri'
        '422':
          description: Your client issued a request that failed validation.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              examples:
                invalid-candidate-not-found-uri:
                  $ref: '#/components/examples/candidate-invalid-not-found-uri'
                rejected-candidate:
                  $ref: '#/components/examples/candidate-scored-too-low'
  /registrations/{id}/score:
    put:
      operationId: scoreSpecification
      summary: Score an existing specification resource
      description: Actualises the score for an existing specification resource.
      tags:
        - registrations
      parameters:
        - name: id
          in: path
          description: ID of specification that needs to be updated
          required: true
          schema:
            type: string
          examples:
            scorecard:
              value: '123'
            invalid-specification-not-found:
              value: not-found
      responses:
        '200':
          description: Ok — existing specification has been scored successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Scorecard'
              examples:
                scorecard:
                  $ref: '#/components/examples/scorecard'
        '400':
          $ref: '#/components/responses/BadRequest'
        '422':
          description: Your client issued a request that failed validation.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              examples:
                invalid-specification-not-found:
                  $ref: '#/components/examples/specification-invalid-not-found'
  /registrations/{id}/component:
    put:
      operationId: implementSpecification
      summary: Implement an existing specification resource
      description: Assigns a component to an existing specification resource, effectively implementing it.
      tags:
        - registrations
      parameters:
        - name: id
          in: path
          description: ID of specification that needs to be updated
          required: true
          schema:
            type: string
          examples:
            valid-component:
              value: '123'
            invalid-specification-not-found:
              value: not-found
            invalid-component-missing-revision:
              value: missing-revision
            invalid-component-malformed-revision:
              value: malformed-revision
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Component'
            examples:
              valid-component:
                $ref: '#/components/examples/valid-component'
              invalid-specification-not-found:
                $ref: '#/components/examples/valid-component'
              invalid-component-missing-revision:
                $ref: '#/components/examples/invalid-component-missing-revision'
              invalid-component-malformed-revision:
                $ref: '#/components/examples/invalid-component-malformed-revision'
      responses:
        '204':
          description: Ok — existing specification has been implemented successfully.
          x-microcks-refs:
            - valid-component
        '400':
          description: The request was malformed or could not be processed.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              examples:
                invalid-component-missing-revision:
                  $ref: '#/components/examples/component-invalid-missing-revision'
                invalid-component-malformed-revision:
                  $ref: '#/components/examples/component-invalid-malformed-revision'
        '422':
          description: Your client issued a request that failed validation.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              examples:
                invalid-specification-not-found:
                  $ref: '#/components/examples/specification-invalid-not-found'
  /registrations/{id}/overlay:
    post:
      operationId: overlaySpecification
      summary: Overlay an existing specification resource
      description: Applies an overlay to an existing specification resource which creates a new revision.
      tags:
        - registrations
      parameters:
        - name: id
          in: path
          description: ID of specification that needs to be updated
          required: true
          schema:
            type: string
          examples:
            valid-overlay:
              value: '123'
            invalid-specification-not-found:
              value: not-found
            invalid-overlay-missing-uri:
              value: missing-uri
            invalid-overlay-malformed-uri:
              value: malformed-uri
            invalid-overlay-not-found:
              value: overlay-not-found
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Candidate'
            examples:
              valid-overlay:
                $ref: '#/components/examples/valid-overlay-uri'
              invalid-specification-not-found:
                $ref: '#/components/examples/valid-overlay-uri'
              invalid-overlay-missing-uri:
                $ref: '#/components/examples/invalid-overlay-missing-uri'
              invalid-overlay-malformed-uri:
                $ref: '#/components/examples/invalid-overlay-malformed-uri'
              invalid-overlay-not-found:
                $ref: '#/components/examples/invalid-overlay-not-found-uri'
      responses:
        '201':
          description: Ok — existing specification has been overlaid successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OverlayApplied'
              examples:
                valid-overlay:
                  $ref: '#/components/examples/specification-overlaid'
        '400':
          description: The request was malformed or could not be processed.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              examples:
                invalid-overlay-missing-uri:
                  $ref: '#/components/examples/overlay-invalid-missing-uri'
                invalid-overlay-malformed-uri:
                  $ref: '#/components/examples/overlay-invalid-malformed-uri'
        '422':
          description: Your client issued a request that failed validation.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              examples:
                invalid-specification-not-found:
                  $ref: '#/components/examples/specification-invalid-not-found'
                invalid-overlay-not-found:
                  $ref: '#/components/examples/overlay-invalid-not-found'
components:
  schemas:
    Candidate:
      type: object
      properties:
        source:
          type: string
          format: uri
          description: URI pointing to the specContent
      required:
        - source
    CandidateProcessed:
      type: object
      description: Response for accepted candidate submission
      properties:
        id:
          type: string
          description: Unique identifier for the candidate
        source:
          type: string
          format: uri
          description: URI pointing to the specContent
        info:
          $ref: '#/components/schemas/Info'
        metadata:
          $ref: '#/components/schemas/Metadata'
        contract:
          $ref: '#/components/schemas/Contract'
        status:
          $ref: '#/components/schemas/Status'
        scorecard:
          $ref: '#/components/schemas/Scorecard'
        violations:
          type: array
          items:
            $ref: '#/components/schemas/Violation'
          description: Detailed violations encountered during processing
      required:
        - source
        - info
        - contract
        - status
    Info:
      type: object
      description: Basic information about the candidate
      properties:
        title:
          type: string
        description:
          type: string
        version:
          type: string
          pattern: ^v\d{1}$
      required:
        - title
        - description
        - version
    Metadata:
      type: object
      description: Additional metadata about the candidate
      properties:
        name:
          type: string
        revision:
          type: string
          pattern: ^\d{1}\.\d{1}\.\d{1}$
        bundleName:
          type: string
        productName:
          type: string
        componentName:
          type: string
        componentRevision:
          type: string
          pattern: ^\d{1}\.\d{1}\.\d{1}$
      required:
        - name
        - revision
        - bundleName
        - productName
    Contract:
      type: string
      description: Specification contract version
      enum:
        - OPENAPI_V30
        - OPENAPI_V31
        - OPENAPI_V32
        - ASYNCAPI_V2
        - ASYNCAPI_V3
        - GRAPHQLS
        - WSDL_V11
        - WSDL_V20
        - OVERLAY_V10
        - OVERLAY_V11
      example: OPENAPI_V31
    Status:
      type: string
      description: Current processing status
      enum:
        - IMPLEMENTED
        - PUBLISHED
        - REGISTERED
        - SCORED
        - VALID
    Scorecard:
      type: object
      description: Detailed scoring breakdown
      properties:
        score:
          $ref: '#/components/schemas/Score'
        grade:
          $ref: '#/components/schemas/Grade'
        dimensions:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
                description: Name of the scoring dimension
              intent:
                type: string
                description: Intent of the scoring dimension
              score:
                $ref: '#/components/schemas/Score'
              grade:
                $ref: '#/components/schemas/Grade'
            required:
              - name
              - intent
              - score
              - grade
        violations:
          type: array
          items:
            $ref: '#/components/schemas/Violation'
          description: Detailed violations encountered during scoring
      required:
        - score
        - grade
    Score:
      type: number
      format: int32
      description: Overall quality score (0-100)
    Grade:
      type: string
      description: Overall quality grade based on score
      enum:
        - A
        - A+
        - A-
        - B
        - B-
        - B-
        - C
        - C+
        - C-
        - D
        - D+
        - D-
    Violation:
      type: object
      description: Violation encountered during processing
      properties:
        code:
          type: string
          description: Code of violation
        detail:
          type: string
          description: Description of violation
        severity:
          type: string
          description: Severity of violation
          enum:
            - MAJOR
            - MINOR
      required:
        - code
        - detail
        - severity
    Component:
      type: object
      description: Component information for implementing a specification
      properties:
        name:
          type: string
          description: Name of the component
        revision:
          type: string
          description: Revision of the component
          pattern: ^\d{1}\.\d{1}\.\d{1}$
      required:
        - name
        - revision
    OverlayApplied:
      type: object
      description: Response for successful overlay application
      properties:
        id:
          type: string
          description: Unique identifier for the overlaid specification
      required:
        - id
...
    Errors:
      type: array
      description: An array of error details to accompany a problem details response.
      maxItems: 1000
      items:
        $ref: '#/components/schemas/ErrorDetail'
    ProblemDetails:
      type: object
      properties:
        type:
          type: string
          description: A URI reference that identifies the problem type.
          format: uri
          maxLength: 1024
        status:
          type: integer
          description: The HTTP status code generated by the origin server for this occurrence of the problem.
          format: int32
          minimum: 100
          maximum: 599
        title:
          type: string
          description: A short, human-readable summary of the problem type. It should not change from occurrence to occurrence of the problem, except for purposes of localization.
          maxLength: 1024
        detail:
          type: string
          description: A human-readable explanation specific to this occurrence of the problem.
          maxLength: 4096
        instance:
          type: string
          description: A URI reference that identifies the specific occurrence of the problem. It may or may not yield further information if dereferenced.
          maxLength: 1024
        code:
          type: string
          description: An API specific error code aiding the provider team understand the error based on their own potential taxonomy or registry.
          maxLength: 50
        errors:
          $ref: '#/components/schemas/Errors'
      required:
        - title
  examples:
    missing-body-property:
      description: Representation of a 400 Bad Request response where a required `body` property was omitted by the consumer.
      value:
        type: https://problems-registry.smartbear.com/missing-body-property
        status: 400
        title: Missing body property
        detail: The request is missing an expected body property.
        code: 400-09
    missing-body-property-with-errors:
      description: Representation of a 400 Bad Request response where a required `body` property was omitted by the consumer.
      value:
        type: https://problems-registry.smartbear.com/missing-body-property
        status: 400
        title: Missing body property
        detail: The request is missing an expected body property.
        code: 400-09
        errors:
          - detail: The body property {name} is required.
            pointer: /name
    missing-request-parameter:
      description: Representation of a 400 Bad Request response where a `query` or `path` parameter was omitted by the consumer.
      value:
        type: https://problems-registry.smartbear.com/missing-request-parameter
        status: 400
        title: Missing request parameter
        detail: The request is missing an expected query or path parameter.
        code: 400-03
...
    missing-request-header:
      description: Representation of a 400 Bad Request response where a `header` parameter was omitted by the consumer.
      value:
        type: https://problems-registry.smartbear.com/missing-request-header
        status: 400
        title: Missing request header
        detail: The request is missing an expected HTTP request header.
        code: 400-02
...
    valid-candidate-uri:
      description: A valid candidate submission example from a local file
      value:
        source: file:///src/test/resources/api/examples/openapi_valid_candidate.yaml
    scored-candidate-uri:
      description: A valid candidate submission example from a local file
      value:
        source: file:///src/test/resources/api/examples/openapi_valid_candidate.yaml
    registered-candidate-uri:
      description: A valid candidate submission example from a local file
      value:
        source: file:///src/test/resources/api/examples/openapi_registered_candidate.yaml
    rejected-candidate-uri:
      description: A valid candidate submission example from a local file
      value:
        source: file:///src/test/resources/api/examples/openapi_rejected_candidate.yaml
    invalid-candidate-missing-uri:
      description: A candidate submission with a missing URI
      value:
        source: ''
    invalid-candidate-malformed-uri:
      description: A candidate submission with a malformed URI
      value:
        source: malformed_uri
    invalid-candidate-not-found-uri:
      description: A candidate submission with a URI that does not exist
      value:
        source: file:///not_found.yaml
    candidate-validated:
      description: A valid candidate submission
      value:
        source: file:///src/test/resources/api/examples/openapi_valid_candidate.yaml
        contract: OPENAPI_V31
        info:
          title: petstore
          description: A sample API for end-to-end test fixtures.
          version: v1
        metadata:
          name: petstore
          revision: 1.0.0
          bundleName: petstore
          productName: platform
        status: VALID
    candidate-scored:
      description: A valid candidate with a quality score
      value:
        source: file:///src/test/resources/api/examples/openapi_scored_candidate.yaml
        contract: OPENAPI_V31
        info:
          title: petstore
          description: A sample API for end-to-end test fixtures.
          version: v1
        metadata:
          name: petstore
          revision: 1.0.0
          bundleName: petstore
          productName: platform
        status: SCORED
        scorecard:
          score: 74
          grade: A-
          dimensions:
            - name: FC - Foundational Compliance
              intent: Base layer of spec validity and structural soundness
              score: 99
              grade: A+
            - name: SEC - Security
              intent: Trust, risk posture, and security compliance
              score: 100
              grade: A+
            - name: DX - Developer Experience
              intent: Clarity, completeness, and ingestion readiness for developers and tooling
              score: 66
              grade: B
            - name: MR - Mock Readiness
              intent: Ease of creating mock servers
              score: 85
              grade: A+
    candidate-registered:
      description: A candidate that has been registered successfully
      value:
        id: '123'
        source: file:///src/test/resources/api/examples/openapi_registered_candidate.yaml
        contract: OPENAPI_V31
        info:
          title: petstore
          description: A sample API for end-to-end test fixtures.
          version: v1
        metadata:
          name: petstore
          revision: 1.0.0
          bundleName: petstore
          productName: platform
        status: REGISTERED
        scorecard:
          score: 74
          grade: A-
          dimensions:
            - name: FC - Foundational Compliance
              intent: Base layer of spec validity and structural soundness
              score: 99
              grade: A+
            - name: SEC - Security
              intent: Trust, risk posture, and security compliance.
              score: 100
              grade: A+
            - name: DX - Developer Experience
              intent: Clarity, completeness, and ingestion readiness for developers and tooling
              score: 66
              grade: B
            - name: ARAX - AI-Readiness & Agent Experience
              intent: Semantic breadth, depth, and agent comprehension for AI systems.
              score: 51
              grade: C-
            - name: AU - Agent Usability
              intent: Functional utility, complexity comfort, and AI orchestration readiness.
              score: 93
              grade: A+
            - name: AID - AI Discoverability
              intent: Findability, semantic richness, and reasoning readiness.
              score: 100
              grade: A+
    candidate-invalid-missing-uri:
      description: A candidate that has been rejected due to missing URI
      value:
        type: https://problems-registry.smartbear.com/validation-error
        status: 400
        title: Validation Error
        detail: The request is invalid.
    candidate-invalid-malformed-uri:
      description: A candidate that has been rejected due to malformed URI
      value:
        type: https://problems-registry.smartbear.com/validation-error
        status: 400
        title: Validation Error
        detail: The request is invalid.
    candidate-invalid-not-found-uri:
      description: A candidate that has been rejected due to URI not found
      value:
        type: https://problems-registry.smartbear.com/validation-error
        status: 422
        title: Validation Error
        detail: The request is invalid.
    candidate-scored-too-low:
      description: A candidate that has been scored but rejected due to low score
      value:
        type: https://problems-registry.smartbear.com/validation-error
        status: 422
        title: Validation Error
        detail: The request is invalid.
    specification-invalid-not-found:
      description: Representation of an 422 Unprocessable Entity response when the specification ID provided does not exist
      value:
        type: https://problems-registry.smartbear.com/validation-error
        status: 422
        title: Validation Error
        detail: The request is invalid.
    scorecard:
      description: A scorecard for a scored candidate
      value:
        score: 74
        grade: A-
        dimensions:
          - name: FC - Foundational Compliance
            intent: Base layer of spec validity and structural soundness
            score: 99
            grade: A+
          - name: SEC - Security
            intent: Trust, risk posture, and security compliance.
            score: 100
            grade: A+
          - name: DX - Developer Experience
            intent: Clarity, completeness, and ingestion readiness for developers and tooling
            score: 66
            grade: B
          - name: ARAX - AI-Readiness & Agent Experience
            intent: Semantic breadth, depth, and agent comprehension for AI systems.
            score: 51
            grade: C-
          - name: AU - Agent Usability
            intent: Functional utility, complexity comfort, and AI orchestration readiness.
            score: 93
            grade: A+
          - name: AID - AI Discoverability
            intent: Findability, semantic richness, and reasoning readiness.
            score: 100
            grade: A+
    valid-component:
      description: A valid component submission example from a local file
      value:
        name: my-component
        revision: 1.0.0
    invalid-component-missing-revision:
      description: An invalid component submission example from a local file
      value:
        name: my-component
    invalid-component-malformed-revision:
      description: An invalid component submission example from a local file
      value:
        name: my-component
        revision: 1.0.0-invalid
    component-invalid-missing-revision:
      description: Representation of a 400 Bad Request response when the component submitted is invalid.
      value:
        type: https://problems-registry.smartbear.com/invalid-body-property-value
        status: 400
        title: Invalid Body Property Value
        detail: The request body contains an invalid body property value.
    component-invalid-malformed-revision:
      description: Representation of a 400 Bad Request response when the component submitted is invalid.
      value:
        type: https://problems-registry.smartbear.com/invalid-body-property-value
        status: 400
        title: Invalid Body Property Value
        detail: The request body contains an invalid body property value.
    valid-overlay-uri:
      description: A valid overlay submission example from a local file
      value:
        source: file:///src/test/resources/api/examples/openapi_valid_overlay.yaml
    invalid-overlay-missing-uri:
      description: An overlay submission with a missing URI
      value:
        source: ''
    invalid-overlay-malformed-uri:
      description: An overlay submission with a malformed URI
      value:
        source: malformed_uri
    invalid-overlay-not-found-uri:
      description: An overlay submission with a URI that does not exist
      value:
        source: file:///not_found.yaml
    overlay-invalid-missing-uri:
      description: An overlay that has been rejected due to missing URI
      value:
        type: https://problems-registry.smartbear.com/bad-request
        status: 400
        title: Validation Error
        detail: The request is invalid.
    overlay-invalid-malformed-uri:
      description: An overlay that has been rejected due to malformed URI
      value:
        type: https://problems-registry.smartbear.com/bad-request
        status: 400
        title: Validation Error
        detail: The request is invalid.
    overlay-invalid-not-found:
      description: An overlay that has been rejected due to URI not found
      value:
        type: https://problems-registry.smartbear.com/validation-error
        status: 422
        title: Validation Error
        detail: The request is invalid.
    specification-overlaid:
      description: A specification that has been overlaid successfully
      value:
        id: '456'
  responses:
    BadRequest:
      description: |
        The request was malformed or could not be processed. 

        Examples of `Bad Request` problem detail responses:

         - [Missing body property](https://problems-registry.smartbear.com/missing-body-property/)
         - [Invalid body property format](https://problems-registry.smartbear.com/invalid-body-property-format/)
         - [Invalid body property value](https://problems-registry.smartbear.com/invalid-body-property-value/)
         - [Missing request header](https://problems-registry.smartbear.com/missing-request-header/)
         - [Invalid request header format](https://problems-registry.smartbear.com/invalid-request-header-format/)
         - [Missing request parameter](https://problems-registry.smartbear.com/missing-request-parameter/)         
         - [Invalid request parameter format](https://problems-registry.smartbear.com/invalid-request-parameter-format/)
         - [Invalid request parameter value](https://problems-registry.smartbear.com/invalid-request-parameter-value/)
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
          examples:
            missingBodyProperty:
              $ref: '#/components/examples/missing-body-property'
            missingBodyPropertyWithErrors:
              $ref: '#/components/examples/missing-body-property-with-errors'
            missingRequestParameter:
              $ref: '#/components/examples/missing-request-parameter'
            missingRequestParameterWithErrors:
              $ref: '#/components/examples/missing-request-parameter-with-errors'
            missingRequestHeader:
              $ref: '#/components/examples/missing-request-header'
            missingRequestHeaderWithErrors:
              $ref: '#/components/examples/missing-request-header-with-errors'
            invalidBodyPropertyFormat:
              $ref: '#/components/examples/invalid-body-property-format'
            invalidBodyPropertyFormatWithErrors:
              $ref: '#/components/examples/invalid-body-property-format-with-errors'
            invalidRequestParameterFormat:
              $ref: '#/components/examples/invalid-request-parameter-format'
            invalidRequestParameterFormatWithErrors:
              $ref: '#/components/examples/invalid-request-parameter-format-with-errors'
            invalidRequestHeaderFormat:
              $ref: '#/components/examples/invalid-request-header-format'
            invalidRequestHeaderFormatWithErrors:
              $ref: '#/components/examples/invalid-request-header-format-with-errors'
            invalidBodyPropertyValue:
              $ref: '#/components/examples/invalid-body-property-value'
            invalidBodyPropertyValueWithErrors:
              $ref: '#/components/examples/invalid-body-property-value-with-errors'
            invalidRequestParameterValue:
              $ref: '#/components/examples/invalid-request-parameter-value'
            invalidRequestParameterValueWithErrors:
              $ref: '#/components/examples/invalid-request-parameter-value-with-errors'